# Remove thread contention from Activity Start/Stop