Fixes #107158 

This change tries to resolve a performance issue related to thread contention when starting and stopping Activities. The new approach is a lock-free, copy-on-write array design which eliminates all thread contention and improves the throughput of Activity.Start/Stop (based on local testing). 

I would like to add some basic tests to the `SynchronizedList<T>` class, but I'm unsure what the policy/procedure is for testing internal classes

I would also like to make sure that this PR has the desired effect of improving request throughput in key benchmarks when telemetry is enabled.


# The problem
The existing code was causing lock contention between request threads that were starting and stopping activities. The approach was to lock and unlock for each item in the collection. So if there were 10 listeners, the request thread would have to acquire and release 10 locks, each of which could contend with other request threads.

# Proposed solution
Use "immutable" arrays of listeners that can be acquired by a request thread without locking. This should make the common/hot path (starting/stopping activities) faster when the system is under load. It should, however, also have a small negative impact on adding/removing listeners, which now becomes a O(N) operation instead of the O(~1) that it was before. However in practice this should have essentially no impact on application performance or allocations

# Local Testing 

My methodology was to start multiple threads (16 vCore machine) and to Start/Stop Activities in a hot loop (worst case).

- 10 worker/request threads
- 50,000 start/dispose activities per thread
- 10 ActivityListeners subscribed up front and not mutated
- 10 runs of benchmark 

## Results Before Changes
Benchmark execution time `00:00:15.1899827`

![image](https://github.com/user-attachments/assets/75a6ec53-f96e-48d2-934e-1901103acc5e)



## Results After Changes
Benchmark execution time `00:00:01.5148120`

![image](https://github.com/user-attachments/assets/cd18bb45-2728-4ebf-9e2e-1c92c4423e74)

## Microbenchmarks
 - With Prometheus metrics exporter 
 - + AddRuntimeInstrumentation
 - 1 synthetic/null tracing/actvity processor
 - 1 synthetic activity listener (to verify invocation counts)
 - 50,000 activity start/stops per thread

Thread Count | Methodology | Max Contentions | Min Contentions | Average Contentions | Average Execution Time (s)
-- | -- | -- | -- | -- | --
1 | Baseline | 0 | 0 | 0 | 0.04198
2 | Baseline | 1 | 0 | 0.125 | 0.05053
4 | Baseline | 9 | 18 | 3.406 | 0.081563
8 | Baseline | 2168 | 1472 | 1743.125 | 0.300675
16 | Baseline | 4436 | 3622 | 4005.375 | 0.6288
1 | New | 0 | 0 | 0 | 0.039686
2 | New | 0 | 0 | 0 | 0.041909
4 | New | 0 | 0 | 0 | 0.047901
8 | New | 0 | 0 | 0 | 0.063951
16 | New | 0 | 0 | 0 | 0.073664

# Open Telemetry Stress test results

## Before

```
opentelemetry-dotnet\test\OpenTelemetry.Tests.Stress.Traces on  main via .NET v8.0.100 🎯 $(TargetFrameworksForTests) took 1m17s
❯ dotnet run --framework net8.0 --configuration Release -- -d 30
Options: {"Concurrency":16,"PrometheusInternalMetricsPort":9464,"DurationSeconds":30}
Run OpenTelemetry.Tests.Stress.Traces.exe --help to see available options.
Running (concurrency = 16, internalPrometheusEndpoint = http://localhost:9464/metrics/), press <Esc> to stop, press <Spacebar> to toggle statistics in the console...
Loops: 41,878,519, Loops/Second: 1,490,815, CPU Cycles/Loop: 10,920, RemainingTime (Seconds): 0
Stopping the stress test...
* Total Running Time (Seconds) 30
* Total Loops: 42,160,488
* Average Loops/Second: 1,404,928
* Average CPU Cycles/Loop: 11,512
* GC Total Allocated Bytes: 20948591184
```

## After
```
opentelemetry-dotnet\test\OpenTelemetry.Tests.Stress.Traces on  main via .NET v8.0.100 🎯 $(TargetFrameworksForTests)
❯ dotnet run --framework net8.0 --configuration Release -- -d 30
Options: {"Concurrency":16,"PrometheusInternalMetricsPort":9464,"DurationSeconds":30}
Run OpenTelemetry.Tests.Stress.Traces.exe --help to see available options.
Running (concurrency = 16, internalPrometheusEndpoint = http://localhost:9464/metrics/), press <Esc> to stop, press <Spacebar> to toggle statistics in the console...
Loops: 451,278,899, Loops/Second: 13,309,562, CPU Cycles/Loop: 1,687, RemainingTime (Seconds): 0
Stopping the stress test...
* Total Running Time (Seconds) 30
* Total Loops: 454,080,963
* Average Loops/Second: 15,048,749
* Average CPU Cycles/Loop: 1,665
* GC Total Allocated Bytes: 225228374680
```