@cijothomas 2024-09-04
> > All those Array.Copy will give the GC some more work to do, I'm curious to know what's the net effect of this change on the affected benchmarks, do you have some numbers already?
> 
> Yes, this will be a performance regression for adding and removing listeners. But with a few caveats
> 
> * Adding/Removing listeners should be relatively infrequent. These allocations should only begin to matter if
>   
>   1. Somone is adding / removing listeners more than a few dozen times per second (which sounds crazy)
>   2. Someone has more than a few dozen listeners and is still churning the list continuously
> * The hot path is broadcasting Activity notifications to listeners, which should be faster in this approach

+1. Adding/Removing listeners is a very rare operation.