@cijothomas 2024-09-04
> @AlgorithmsAreCool If you'd like to, you can run the stress test tool built by OpenTelemetry for testing OTel itself! https://github.com/open-telemetry/opentelemetry-dotnet/tree/main/test/OpenTelemetry.Tests.Stress.Traces - The stress test shows the throughput, and it should show an improvement, when applying changes from this PR.

In my dev box (16 logical cores), the stress test for Traces is showing a throughput of ~2M/sec. Whereas the one to Logs show 50M/sec. A good chunk of this difference is likely the contention being fixed here. The stress test show 0 contention for Logs as Logs has no feature of "listeners"! Hopefully, with this change, Traces should see a huge jump in throughput!