@tommcdon 2024-09-05
> @dotnet/area-system-diagnostics-activity I'm not sure how to proceed, the build failures as far as i can tell do not seem to be related

Adding @elinor-fung @jkoritzinsky as the failures seem to be COM/binder source gen related