@Kielek 2024-09-06
@tarekgh, based on what I see it is/will be targeted to .NET10. Do you have also plan to _backport_ to .NET9 or it is to late?