@AlgorithmsAreCool 2024-09-06
> @Kielek
> 
> > based on what I see it is/will be targeted to .NET10. Do you have also plan to backport to .NET9 or it is to late?
> 
> From previous discussions parties reported this were ok to have in next release 10. Our current plan is to have it in .NET 10 to reduce any risk for .NET 9.0. Is it blocking any real scenario for you? Please elaborate more why you are asking to have the fix in 9.0.

For my team, it isn't necessarily blocking our use of tracing, but we operate at least 1 service that is very low latency that might notice this change.

I'm curious how much of the 15% impact to topline throughput this change wins back.