@tarekgh 2024-10-29
/backport to 9.0-staging